fix(net): propiedades de item en float64, y el avatar remoto con pantalón y calzado de trabajo
Propiedades de item (ADR-072): el cliente las escribía como float32 y Garment Zones empaqueta
32 bits; float32 solo es exacto hasta 2^24, así que la rotura de las zonas 6-7 (los
antebrazos de la chaqueta) se corrompía en report_inventory, guardado y cadáveres. Ahora
MsgPackWriter.WriteDouble / MsgPackReader.ReadDouble / IPCParse.ToDouble; el backend ya las
guardaba en f64. Sin cambio de campos ni de esquema. Test de ida y vuelta con las 8 zonas.

Avatar remoto (ADR-149 R4a, sin wire): el pantalón y el calzado de trabajo ya viajaban en
equipment pero el guardarropa del proxy no los conocía y SetClothing dejaba el hueco vacío.
RemoteAvatarPrefabBuilder.WireBackroomsGarments los añade con BackroomsGarmentVisualsBuilder
.WireRemoteAvatar (bucle de prendas separado del muñeco, sin la de encima ni componentes de UI).
EnsureMaterial pone siempre el shader de rotura: el material es compartido y el donante del
proxy es del vendor (el primer re-horneado lo cambiaba a URP/Lit). Re-horneado en una pasada:
_controller intacto, RealForm presente, prefab solo con inserciones.

EditMode: 135/135 (ropa, rotura, piel, mangas, msgpack, restaurador, cadáveres, proxy).

+++ Assets/Scripts/Network/IPCClient.cs
@@ -782,7 +782,8 @@ namespace BackroomsSurvival.Net
             {
                 w.WriteMapHeader(2);
                 w.WriteString("id"); w.WriteInt(props[p].id);
-                w.WriteString("value"); w.WriteFloat((float)props[p].value);
+                // float64: la rotura de una prenda empaqueta 32 bits y float32 solo es exacto hasta 2^24 (el backend guarda f64).
+                w.WriteString("value"); w.WriteDouble(props[p].value);
             }
         }
